Extract `createEmptyProject` to own file
